requirements.txt
boto3==1.10.28
botocore==1.13.28
cloudpickle==1.2.2
cycler==0.10.0
docutils==0.15.2
future==0.18.2
gym==0.15.4
gym-recording==0.0.1
imageio==2.6.1
jmespath==0.9.4
kiwisolver==1.1.0
matplotlib==3.1.2
numpy==1.17.4
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==6.2.1
PyBrain==0.3
pygame==1.9.6
pyglet==1.3.2
pyparsing==2.4.5
pypng==0.0.20
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
scipy==1.3.3
six==1.13.0
urllib3==1.25.7
-e git+https://github.com/lburger98/py-vgdl.git@04026dbea780ab3f0c2e7baf46f8b04c5c531f86#egg=vgdl
